[cleanup] Cleanup + version bumped to 0.7-SNAPSHOT
