+++ src/main/scala/AndroidSDK.scala
@@ -3,9 +3,10 @@ package org.scalasbt.androidplugin
 import sbt._
 
 import Keys._
-import AndroidKeys._
+import AndroidPlugin._
+import AndroidHelpers._
 
-object AndroidDefaults {
+object AndroidSDK {
   val DefaultAaaptName = "aapt"
   val DefaultAadbName = "adb"
   val DefaultAaidlName = "aidl"
@@ -20,7 +21,8 @@ object AndroidDefaults {
   val DefaultManifestSchema = "http://schemas.android.com/apk/res/android"
   val DefaultEnvs = List("ANDROID_SDK_HOME", "ANDROID_SDK_ROOT", "ANDROID_HOME")
 
-  lazy val settings: Seq[Setting[_]] = Seq (
+  lazy val settings: Seq[Setting[_]] = inConfig(Android) {Seq(
+    // Command executable names
     aaptName := DefaultAaaptName,
     adbName := DefaultAadbName,
     aidlName := DefaultAaidlName,
@@ -34,7 +36,22 @@ object AndroidDefaults {
     resourcesApkName := DefaultResourcesApkName,
     manifestSchema := DefaultManifestSchema,
     envs := DefaultEnvs,
-    // a list of modules which are already included in Android
+
+    // On Windows, `dx` is a .bat executable, but not on Linux
+    osDxName <<= (dxName) (_ + osBatchSuffix),
+
+    // Find the path of the Android SDK
+    sdkPath <<= (envs, baseDirectory) ((e, d) => determineAndroidSdkPath(e, d)),
+
+    // Generate the full paths of the SDK tools
+    toolsPath <<= (sdkPath) (_ / "tools"),
+    dbPath <<= (platformToolsPath, adbName) (_ / _),
+    platformToolsPath <<= (sdkPath) (_ / "platform-tools"),
+    aaptPath <<= (platformToolsPath, aaptName) (_ / _),
+    idlPath <<= (platformToolsPath, aidlName) (_ / _),
+    dxPath <<= (platformToolsPath, osDxName) (_ / _),
+
+    // A list of modules which are already included in Android
     preinstalledModules := Seq[ModuleID](
       ModuleID("org.apache.httpcomponents", "httpcore", null),
       ModuleID("org.apache.httpcomponents", "httpclient", null),
@@ -42,5 +59,5 @@ object AndroidDefaults {
       ModuleID("commons-logging", "commons-logging", null),
       ModuleID("commons-codec", "commons-codec", null)
     )
-  )
+  )}
 }
