+++ src/main/scala/PasswordManager.scala
@@ -3,7 +3,7 @@ package org.scalasbt.androidplugin
 import sbt._
 
 import Keys._
-import AndroidKeys._
+import AndroidPlugin._
 
 object PasswordManager extends PWManager {
   lazy val settings: Seq[Setting[_]] = inConfig(Android) (Seq (
