+++ src/main/scala/legacy/Github.scala
@@ -1,4 +1,4 @@
-package org.scalasbt.androidplugin
+package org.scalasbt.androidplugin.legacy
 
 import java.io.{File,FileInputStream,OutputStreamWriter,PrintWriter}
 import java.net.{URL, HttpURLConnection}
@@ -8,12 +8,22 @@ import scala.xml.Node
 import sbt._
 import Keys._
 
-import AndroidKeys._
+import org.scalasbt.androidplugin.AndroidPlugin._
+import org.scalasbt.androidplugin.PasswordManager
 
 object Github {
   val gitConfig = new File(System.getenv("HOME"), ".gitconfig")
   val apkMime = "application/vnd.android.package-archive"
   val gitDownloads = "https://api.github.com/repos/%s/downloads"
+
+  /** Github keys **/
+  object Keys {
+    val uploadGithub = TaskKey[Option[String]]("github-upload", "Upload file to github")
+    val deleteGithub = TaskKey[Unit]("github-delete", "Delete file from github")
+    val githubRepo   = SettingKey[String]("github-repo", "Github repo")
+  }
+
+  import Keys._
 
   lazy val settings: Seq[Setting[_]] = inConfig(Android) (Seq (
     uploadGithub <<= (prepareMarket, githubRepo, cachePasswords, streams) map { (path, repo, cache, s) =>
